+++ libs/application/templates/european-health-insurance-card/src/forms/EuropeanHealthInsuranceCardForm.tsx
@@ -5,6 +5,10 @@ import {
   NationalRegistrySpouseApi,
   NationalRegistryUserApi,
 } from '@island.is/application/types'
+import {
+  EhicApplyForPhysicalCardApi,
+  EhicCardResponseApi,
+} from '../dataProviders'
 import { Form, FormModes } from '@island.is/application/types'
 import {
   buildCheckboxField,
@@ -18,15 +22,16 @@ import {
   buildSubmitField,
 } from '@island.is/application/core'
 
-import { EhicCardResponseApi } from '../dataProviders'
 import { NationalRegistry } from '../lib/types'
 import { europeanHealthInsuranceCardApplicationMessages as e } from '../lib/messages'
 
-export const EuropeanHealthInsuranceCardApplyPlastic: Form = buildForm({
-  id: 'EuropeanHealthInsurancePrerequisitiesForm',
+/* eslint-disable-next-line */
+export interface EuropeanHealthInsuranceCardProps {}
+
+export const EuropeanHealthInsuranceCardForm: Form = buildForm({
+  id: 'EuropeanHealthInsuranceCardForm',
   title: '',
   mode: FormModes.DRAFT,
-  renderLastScreenButton: false,
   children: [
     buildSection({
       id: 'intro',
@@ -49,13 +54,37 @@ export const EuropeanHealthInsuranceCardApplyPlastic: Form = buildForm({
     buildSection({
       id: 'data',
       title: e.data.sectionLabel,
-      children: [],
-    }),
-
-    buildSection({
-      id: 'plastic',
-      title: e.applicants.sectionTitle,
       children: [
+        buildExternalDataProvider({
+          title: e.data.sectionTitle,
+          checkboxLabel: e.data.dataCollectionCheckboxLabel,
+          id: 'approveExternalData',
+          description: '',
+          dataProviders: [
+            buildDataProviderItem({
+              provider: NationalRegistryUserApi,
+              title: 'Þjóðskrá Íslands',
+              subTitle:
+                'Við þurfum að sækja þessi gögn úr þjóðskrá. Lögheimili, hjúskaparstaða, maki og afkvæmi.',
+            }),
+            buildDataProviderItem({
+              provider: NationalRegistrySpouseApi,
+              title: '',
+              subTitle: '',
+            }),
+            buildDataProviderItem({
+              provider: ChildrenCustodyInformationApi,
+              title: '',
+              subTitle: '',
+            }),
+            buildDataProviderItem({
+              provider: EhicCardResponseApi,
+              title: 'Sjúkratryggingar',
+              subTitle:
+                'Upplýsingar um stöðu heimildar á evrópska sjúktryggingakortinu',
+            }),
+          ],
+        }),
         buildMultiField({
           id: 'plastic',
           title: e.applicants.sectionTitle,
@@ -104,15 +133,57 @@ export const EuropeanHealthInsuranceCardApplyPlastic: Form = buildForm({
                 return applying as Array<{ value: any; label: string }>
               },
             }),
+          ],
+        }),
+
+        // Has to be here so that the submit button appears (does not appear if no screen is left).
+        // Tackle that as AS task.
+        // buildDescriptionField({
+        //   id: 'unused',
+        //   title: '',
+        //   description: '',
+        // }),
+      ],
+    }),
+
+    buildSection({
+      id: 'temp',
+      title: e.temp.sectionLabel,
+      children: [
+        buildMultiField({
+          id: 'tempApplicants',
+          title: e.temp.sectionTitle,
+          description: e.temp.sectionDescription,
+          children: [
+            buildCheckboxField({
+              id: 'applyForPDF',
+              backgroundColor: 'white',
+              title: '',
+              options: (application: Application) => {
+                const applying = []
+                console.log(application)
+                const ans = application.answers.applyForPlastic as Array<any>
+                console.log('answers')
+                console.log(ans)
+                for (const i in ans) {
+                  console.log([ans[i][0], ans[i][1]])
+                  applying.push({
+                    value: [ans[i][0], ans[i][1]],
+                    label: ans[i][1],
+                  })
+                }
+                return applying as Array<{ value: any; label: string }>
+              },
+            }),
             buildSubmitField({
-              id: 'submit',
+              id: 'submit-pdf',
               title: e.review.submitButtonLabel,
               refetchApplicationAfterSubmit: true,
               placement: 'footer',
               actions: [
                 {
                   event: DefaultEvents.SUBMIT,
-                  name: e.applicants.submitButtonLabel,
+                  name: e.temp.submitButtonLabel,
                   type: 'primary',
                 },
               ],
@@ -122,17 +193,11 @@ export const EuropeanHealthInsuranceCardApplyPlastic: Form = buildForm({
         // Has to be here so that the submit button appears (does not appear if no screen is left).
         // Tackle that as AS task.
         buildDescriptionField({
-          id: 'unused5',
+          id: 'pdf-Unused',
           title: '',
           description: '',
         }),
       ],
-    }),
-
-    buildSection({
-      id: 'temp',
-      title: e.temp.sectionLabel,
-      children: [],
     }),
 
     buildSection({
@@ -148,3 +213,5 @@ export const EuropeanHealthInsuranceCardApplyPlastic: Form = buildForm({
     }),
   ],
 })
+
+export default EuropeanHealthInsuranceCardForm
